+++ src/app/page.js
@@ -1,0 +1,10 @@
+import { Home } from '../pages/Home'
+
+export const metadata = {
+  title: 'Mark Li - Software Developer & Amateur Musician',
+  description: 'Personal website of Mark Li, software developer and amateur musician',
+}
+
+export default function HomePage() {
+  return <Home />
+}