+++ src/components/Navbar/mobile/MobileNavbarButton.js
@@ -1,11 +1,11 @@
 import React from "react";
-import { Link } from "react-router-dom";
+import Link from "next/link";
 
 export function MobileNavbarButton({ title, link, toggle }) {
   return (
     <div>
       <Link
-        to={link}
+        href={link}
         onClick={toggle}
         className="flex items-center justify-center py-2 hover:bg-slate-500 font-light rounded-md"
       >