+++ tailwind.config.js
@@ -1,13 +1,18 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
-  content: ["./src/**/*.{js,jsx,ts,tsx}"],
+  content: [
+    './src/pages/**/*.{js,ts,jsx,tsx,mdx}',
+    './src/components/**/*.{js,ts,jsx,tsx,mdx}',
+    './src/app/**/*.{js,ts,jsx,tsx,mdx}',
+    './src/projects/**/*.{js,ts,jsx,tsx,mdx}',
+  ],
   theme: {
     extend: {
       fontFamily: {
-        inter: ["Inter", "serif"],
-        source: ["Source", "serif"],
+        inter: ['Inter', 'serif'],
+        source: ['Source Code Pro', 'serif'],
       },
     },
   },
   plugins: [],
-};
+}